+++ app/Http/Controllers/RealtorController.php
@@ -87,6 +87,10 @@ class RealtorController extends Controller
      */
     public function edit(Realtor $realtor)
     {
+        if (!Gate::allows('update-realtor', $realtor)) {
+            abort(403);
+        }
+
         return view('realtors.edit',compact('realtor'));
     }
     
@@ -100,6 +104,10 @@ class RealtorController extends Controller
 
     public function update(Request $request, Realtor $realtor)
     {
+        if (!Gate::allows('update-realtor', $realtor)) {
+            abort(403);
+        }
+
         $request->validate([
             'phone' => 'required|regex:/^([0-9\s\-\+\(\)]*)$/|min:10',
         ]);
@@ -118,6 +126,10 @@ class RealtorController extends Controller
      */
     public function destroy(Realtor $realtor)
     {
+        if (!Gate::allows('delete-realtor', $realtor)) {
+            abort(403);
+        }
+
         $realtor->delete();
     
         return redirect()->route('realtors.index')
